+++ codegen-cli/src/test/java/com/networknt/codegen/YAMLFileParameterizerTest.java
@@ -1,8 +1,6 @@
 package com.networknt.codegen;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import org.junit.Ignore;
-import org.junit.Test;
 
 import java.io.File;
 import java.io.FilenameFilter;
@@ -13,9 +11,11 @@ import java.util.Map;
 import java.util.Set;
 
 import static java.io.File.separator;
-import static org.junit.Assert.assertTrue;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
 
-@Ignore
+@Disabled
 public class YAMLFileParameterizerTest {
 	private static String configName = "/config.json";