+++ codegen-web/src/test/java/com/networknt/codegen/handler/GeneratorServiceHandlerTest.java
@@ -10,9 +10,9 @@ import io.undertow.client.ClientRequest;
 import io.undertow.client.ClientResponse;
 import io.undertow.util.Headers;
 import io.undertow.util.Methods;
-import org.junit.Assert;
-import org.junit.ClassRule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import static org.junit.jupiter.api.Assertions.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xnio.IoUtils;
@@ -30,10 +30,10 @@ import java.util.concurrent.atomic.AtomicReference;
 /**
  * Created by steve on 25/04/17.
  */
+@ExtendWith(TestServer.class)
 public class GeneratorServiceHandlerTest {
     private static String auth = "Bearer eyJraWQiOiIxMDAiLCJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJ1cm46Y29tOm5ldHdvcmtudDpvYXV0aDI6djEiLCJhdWQiOiJ1cm46Y29tLm5ldHdvcmtudCIsImV4cCI6MTgwOTAxMTkyMCwianRpIjoiR2NpMGJQZXoxb0hxT1VzYUZ6WmRkdyIsImlhdCI6MTQ5MzY1MTkyMCwibmJmIjoxNDkzNjUxODAwLCJ2ZXJzaW9uIjoiMS4wIiwidXNlcl9pZCI6InN0ZXZlIiwidXNlcl90eXBlIjoiRU1QTE9ZRUUiLCJjbGllbnRfaWQiOiJmN2Q0MjM0OC1jNjQ3LTRlZmItYTUyZC00YzU3ODc0MjFlNzIiLCJzY29wZSI6WyJjb2RlZ2VuLnIiLCJjb2RlZ2VuLnciLCJzZXJ2ZXIuaW5mby5yIl19.MMIjxGlQknwtlizh80wX1oB75N8wfhqMttP7i3mpKwBa-zUKZcjgtE4rmc39qYXPti9ge3uGHWCQdMOlimf4Psoah-qtsQmZhuCOwejN_OhwvlIbLxCYYP9WaQM_zyuDv6luFO5ETsZQ0-QkxQHkBq1Y3D0VpHNuNlN3ulK5sK678XKw_2VNP6JM85hh1QW8pWYappkAledvyfJC3w5sUWdi3kP_rTGiumXYA0ZoY-hdp9erlin0ClEZ7qPmHSHW_TFjuRr_6rDwE1Xg-U6wNk1hMMFF4161nfdPIhDhCsjG3J4hM0y0ZluI3BnQaGKwtqyDNIJGiiTkK1ckX1qaIw";
 
-    @ClassRule
     public static TestServer server = TestServer.getInstance();
     static final boolean enableHttp2 = server.getServerConfig().isEnableHttp2();
     static final boolean enableHttps = server.getServerConfig().isEnableHttps();
@@ -100,9 +100,9 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(400, statusCode);
+        assertEquals(400, statusCode);
         if(statusCode == 400) {
-            Assert.assertTrue(body.contains("ERR11200"));
+            assertTrue(body.contains("ERR11200"));
         }
     }
 
@@ -152,9 +152,9 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(400, statusCode);
+        assertEquals(400, statusCode);
         if(statusCode == 400) {
-            Assert.assertTrue(body.contains("ERR11100"));
+            assertTrue(body.contains("ERR11100"));
         }
     }
 
@@ -204,9 +204,9 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(400, statusCode);
+        assertEquals(400, statusCode);
         if(statusCode == 400) {
-            Assert.assertTrue(body.contains("ERR11100"));
+            assertTrue(body.contains("ERR11100"));
         }
     }
 
@@ -250,8 +250,8 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(200, statusCode);
-        Assert.assertNotNull(body);
+        assertEquals(200, statusCode);
+        assertNotNull(body);
     }
 
     /**
@@ -294,8 +294,8 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(200, statusCode);
-        Assert.assertNotNull(body);
+        assertEquals(200, statusCode);
+        assertNotNull(body);
     }
 
     /**
@@ -338,8 +338,8 @@ public class GeneratorServiceHandlerTest {
         }
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-        Assert.assertEquals(200, statusCode);
-        Assert.assertNotNull(body);
+        assertEquals(200, statusCode);
+        assertNotNull(body);
     }
 
     /**
@@ -383,8 +383,8 @@ public class GeneratorServiceHandlerTest {
         int statusCode = reference.get().getResponseCode();
         String body = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
         System.out.println(body);
-        Assert.assertEquals(200, statusCode);
-        Assert.assertNotNull(body);
+        assertEquals(200, statusCode);
+        assertNotNull(body);
     }
 
 }