+++ codegen-web/src/test/java/com/networknt/codegen/handler/TestServer.java
@@ -2,36 +2,36 @@ package com.networknt.codegen.handler;
 
 import com.networknt.server.Server;
 import com.networknt.server.ServerConfig;
-import org.junit.rules.ExternalResource;
+import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 import java.util.concurrent.atomic.AtomicInteger;
 
-/**
- * Created by steve on 25/04/17.
- */
-public class TestServer extends ExternalResource {
+public class TestServer implements BeforeAllCallback, AfterAllCallback {
     static final Logger logger = LoggerFactory.getLogger(TestServer.class);
 
     private static final AtomicInteger refCount = new AtomicInteger(0);
-    private static Server server;
 
     private static final TestServer instance  = new TestServer();
+
+    private TestServer() {
+        logger.info("TestServer is constructed!");
+    }
 
     public static TestServer getInstance () {
         return instance;
     }
 
+
     public ServerConfig getServerConfig() {
         return Server.getServerConfig();
     }
 
-    private TestServer() {
-
-    }
-
     @Override
-    protected void before() {
+    public void beforeAll(ExtensionContext context) throws Exception {
         try {
             if (refCount.get() == 0) {
                 Server.start();
@@ -40,13 +40,21 @@ public class TestServer extends ExternalResource {
         finally {
             refCount.getAndIncrement();
         }
+
+        Runtime.getRuntime().addShutdownHook(new Thread(() -> clean()));
     }
 
     @Override
-    protected void after() {
+    public void afterAll(ExtensionContext context) throws Exception {
+        clean();
+    }
+
+    protected void clean() {
         refCount.getAndDecrement();
         if (refCount.get() == 0) {
             Server.stop();
         }
+
+
     }
 }