+++ light-rest-4j/src/test/java/com/networknt/codegen/ConfigTest.java
@@ -1,9 +1,10 @@
 package com.networknt.codegen;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeAll;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.IOException;
 
@@ -13,7 +14,7 @@ public class ConfigTest {
     public static JsonNode anyConfig = null;
     public static JsonNode anyYamlConfig = null;
 
-    @BeforeClass
+    @BeforeAll
     public static void setUp() throws IOException {
         // load config file
         anyConfig = Generator.jsonMapper.readTree(OpenApiLightGeneratorTest.class.getResourceAsStream(configName));
@@ -22,17 +23,17 @@ public class ConfigTest {
 
     @Test
     public void testDbName() {
-        Assert.assertEquals("mysql", anyConfig.path("dbInfo").path("name").textValue());
+        assertEquals("mysql", anyConfig.path("dbInfo").path("name").textValue());
     }
 
     @Test
     public void testDbSupport() {
-        Assert.assertTrue(anyConfig.get("supportDb").booleanValue());
+        assertTrue(anyConfig.get("supportDb").booleanValue());
     }
 
     @Test
     public void testUseLightProxy() {
-        Assert.assertTrue(anyYamlConfig.get("useLightProxy").booleanValue());
+        assertTrue(anyYamlConfig.get("useLightProxy").booleanValue());
     }
 
 }